+++ tests/unittests/test_completers.py
@@ -31,7 +31,7 @@ def test_LUF_completer_touch_words():
 
 def test_newbie_mode_complete_without_meta_dict():
     fake_document = MagicMock()
-    fake_document.text_before_cursor = "GEOR"
+    fake_document.text_before_cursor = fake_document.text = "GEOR"
     completer = IRedisCompleter(hint=False)
     completions = list(completer.get_completions(fake_document, None))
     assert [word.text for word in completions] == ["GEORADIUS", "GEORADIUSBYMEMBER"]
@@ -43,7 +43,7 @@ def test_newbie_mode_complete_without_meta_dict():
 
 def test_newbie_mode_complete_with_meta_dict():
     fake_document = MagicMock()
-    fake_document.text_before_cursor = "GEOR"
+    fake_document.text_before_cursor = fake_document.text = "GEOR"
     completer = IRedisCompleter(hint=True)
     completions = list(completer.get_completions(fake_document, None))
 
@@ -69,7 +69,7 @@ def test_newbie_mode_complete_with_meta_dict():
 
 def test_newbie_mode_complete_with_meta_dict_command_is_lowercase():
     fake_document = MagicMock()
-    fake_document.text_before_cursor = "geor"
+    fake_document.text_before_cursor = fake_document.text = "geor"
     completer = IRedisCompleter(hint=True)
     completions = list(completer.get_completions(fake_document, None))