+++ iredis/exceptions.py
@@ -1,10 +1,18 @@
-class UsageError(Exception):
+class IRedisException(Exception):
     pass
 
 
-class InvalidArguments(UsageError):
+class UsageError(IRedisException):
+    pass
+
+
+class InvalidArguments(IRedisException):
     """Invalid argument(s)"""
 
 
-class NotRedisCommand(UsageError):
+class NotRedisCommand(IRedisException):
     """Not a Redis command"""
+
+
+class AmbiguousCommand(IRedisException):
+    """Command is not finished, don't it's command's name"""